Fix example errors for CrossNamespacePodAffinity
Remove references to CrossNamespaceAffinity

The scope CrossNamespaceAffinity does not exist. Attempting to feed the example
YAML to `kubectl create` results in the following error:

> The ResourceQuota "disable-cross-namespace-affinity" is invalid:
> * spec.scopeSelector.matchExpressions.scopeName: Invalid value:
>  "CrossNamespaceAffinity": unsupported scope

Add missing operator for CrossNamespacePodAffinity

Trying to create the example ResourceQuotas without an operator results in the
following error from `kubectl create`:

> The ResourceQuota "disable-cross-namespace-affinity" is invalid:
> * spec.scopeSelector.matchExpressions.operator: Invalid value: "": must be
>   'Exist' when scope is any of ResourceQuotaScopeTerminating,
>   ResourceQuotaScopeNotTerminating, ResourceQuotaScopeBestEffort,
>   ResourceQuotaScopeNotBestEffort or
>   ResourceQuotaScopeCrossNamespacePodAffinity
> * spec.scopeSelector.matchExpressions.operator: Invalid value: "": not a valid
>     selector operator

The error message itself has another bug, as the operator is Exist*s*, not
Exist.

Signed-off-by: Simon Engmann <simon.engmann@sovity.de>