Revise naming style for feature state files
The current naming style creates some problems for tracking/maintaining the list
of feature gates. People gets confused about how to convert the feature name
to the kebal-case. The sorted list of the converted names are different from the
output of the help of command line tools. This further complicates the cross-checking
of missing/removed features.

This PR changes the style to be a more straigth-forward way, i.e. use
the feature name as the file name. No extra conversion should be included.
This makes the lives of everyone a lot easier.